refactoring moves launch and exec script creation to executor base
