+++ src/radical/pilot/agent/executing/popen.py
@@ -4,7 +4,6 @@ __copyright__ = 'Copyright 2013-2022, The RADICAL-Cybertools Team'
 __license__   = 'MIT'
 
 import os
-import stat
 import time
 import queue
 import atexit
@@ -48,9 +47,6 @@ signal.signal(signal.SIGINT,  _kill)
 #
 class Popen(AgentExecutingComponent):
 
-    _header    = '#!/bin/sh\n'
-    _separator = '\n# ' + '-' * 78 + '\n'
-
     # flags for the watcher queue
     TO_WATCH  = 0
     TO_CANCEL = 1
@@ -64,14 +60,10 @@ class Popen(AgentExecutingComponent):
 
         self._watch_queue = queue.Queue()
 
-        self._pid = self.session.cfg.pid
-
         # run watcher thread
         self._watcher = mt.Thread(target=self._watch)
-      # self._watcher.daemon = True
+        self._watcher.daemon = True
         self._watcher.start()
-
-      # self._log.debug('popen initialize stop')
 
 
     # --------------------------------------------------------------------------
@@ -206,168 +198,14 @@ class Popen(AgentExecutingComponent):
 
         launcher = self._rm.find_launcher(task)
 
-        self._create_launch_script(launcher, task)
-        self._create_exec_script(launcher, task)
-        self._launch_task(task)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _create_launch_script(self, launcher, task):
+        exec_path   = self._create_exec_script(launcher, task)
+        launch_path = self._create_launch_script(launcher, task, exec_path)
 
         tid  = task['uid']
-        td   = task['description']
         sbox = task['task_sandbox_path']
-
-        if not launcher:
-            raise RuntimeError('no launcher found for task %s' % tid)
-
-        self._log.debug('Launching task with %s', launcher.name)
-
-        launch_script = '%s.launch.sh'        % tid
-        exec_script   = '%s.exec.sh'          % tid
-        exec_path     = '$RP_TASK_SANDBOX/%s' % exec_script
-
-        ru.rec_makedir(sbox)
-
-        with ru.ru_open('%s/%s' % (sbox, launch_script), 'w') as fout:
-
-            tmp  = ''
-            tmp += self._header
-            tmp += self._separator
-            tmp += self._get_rp_env(task)
-            tmp += self._get_rp_funcs()
-            tmp += self._separator
-            tmp += self._get_prof('launch_start', tid)
-
-            tmp += self._separator
-            tmp += '# change to task sandbox\n'
-            tmp += 'cd $RP_TASK_SANDBOX\n'
-
-            tmp += self._separator
-            tmp += '# prepare launcher env\n'
-            tmp += self._get_launch_env(launcher)
-
-            tmp += self._separator
-            tmp += '# pre-launch commands\n'
-            tmp += self._get_prof('launch_pre', tid)
-            tmp += self._get_prep_launch(task, sig='pre_launch')
-
-            tmp += self._separator
-            tmp += '# launch commands\n'
-            tmp += self._get_prof('launch_submit', tid)
-            tmp += self._get_launch(task, launcher, exec_path)
-            tmp += self._get_prof('launch_collect', tid,
-                                  msg='RP_LAUNCH_PID=$RP_LAUNCH_PID')
-
-            tmp += self._separator
-            tmp += '# post-launch commands\n'
-            tmp += self._get_prof('launch_post', tid)
-            tmp += self._get_prep_launch(task, sig='post_launch')
-
-            tmp += self._separator
-            tmp += self._get_prof('launch_stop', tid)
-            tmp += 'exit $RP_RET\n'
-
-            tmp += self._separator
-            tmp += '\n'
-
-            fout.write(tmp)
-
-    # --------------------------------------------------------------------------
-    #
-    def _create_exec_script(self, launcher, task):
-
-        tid  = task['uid']
-        td   = task['description']
-        sbox = task['task_sandbox_path']
-
-        exec_script   = '%s.exec.sh'   % tid
-        launch_script = '%s.launch.sh' % tid
-
-        # the exec shell script runs the same set of commands for all ranks.
-        # However, if the ranks need different GPU's assigned, or if either pre-
-        # or post-exec directives contain per-rank dictionaries, then we switch
-        # per-rank in the script for all sections between pre- and post-exec.
-
-        n_ranks = td['ranks']
-        slots   = task.setdefault('slots', {})
-
-        self._extend_pre_exec(td, slots.get('ranks'))
-
-        with ru.ru_open('%s/%s' % (sbox, exec_script), 'w') as fout:
-
-            tmp  = ''
-            tmp += self._header
-            tmp += self._separator
-            tmp += self._get_rp_env(task)
-            tmp += self._get_rp_funcs()
-            tmp += self._separator
-            tmp += '# rank ID\n'
-            tmp += self._get_rank_ids(n_ranks, launcher)
-            tmp += self._separator
-            tmp += self._get_prof('exec_start', tid)
-
-            tmp += self._get_task_env(task, launcher)
-
-            tmp += self._separator
-            tmp += '# pre-exec commands\n'
-            tmp += self._get_prof('exec_pre', tid)
-            tmp += self._get_prep_exec(task, n_ranks, sig='pre_exec')
-
-            tmp += self._separator
-            tmp += '# execute rank\n'
-            tmp += self._get_prof('rank_start', tid)
-            tmp += self._get_exec(task, launcher)
-            tmp += self._get_prof('rank_stop', tid,
-                                  msg='RP_EXEC_PID=$RP_EXEC_PID:'
-                                      'RP_RANK_PID=$RP_RANK_PID')
-
-            tmp += self._separator
-            tmp += '# post-exec commands\n'
-            tmp += self._get_prof('exec_post', tid)
-            tmp += self._get_prep_exec(task, n_ranks, sig='post_exec')
-
-            tmp += self._separator
-            tmp += self._get_prof('exec_stop', tid)
-            tmp += 'exit $RP_RET\n'
-
-            tmp += self._separator
-            tmp += '\n'
-
-            fout.write(tmp)
-
-
-        # make sure scripts are executable
-        st_l = os.stat('%s/%s' % (sbox, launch_script))
-        st_e = os.stat('%s/%s' % (sbox, exec_script))
-        os.chmod('%s/%s' % (sbox, launch_script), st_l.st_mode | stat.S_IEXEC)
-        os.chmod('%s/%s' % (sbox, exec_script),   st_e.st_mode | stat.S_IEXEC)
-
-        # make sure the sandbox exists
-        self._prof.prof('task_mkdir', uid=tid)
-        ru.rec_makedir(sbox)
-        self._prof.prof('task_mkdir_done', uid=tid)
-
-        # need to set `DEBUG_5` or higher to get slot debug logs
-        if self._log._debug_level >= 5:
-            ru.write_json('%s/%s.sl' % (sbox, tid), slots)
-
-    # --------------------------------------------------------------------------
-    #
-    def _launch_task(self, task):
-
-        tid  = task['uid']
-        td   = task['description']
-        sbox = task['task_sandbox_path']
-
-        exec_script   = '%s.exec.sh'   % tid
-        launch_script = '%s.launch.sh' % tid
 
         # launch and exec script are done, get ready for execution.
-        cmdline = '%s/%s' % (sbox, launch_script)
-
-        self._log.info('Launching task %s via %s in %s', tid, cmdline, sbox)
+        self._log.info('Launching task %s via %s in %s', tid, launch_path, sbox)
 
         _launch_out_h = ru.ru_open('%s/%s.launch.out' % (sbox, tid), 'w')
 
@@ -377,7 +215,7 @@ class Popen(AgentExecutingComponent):
         _start_new_session = self.session.rcfg.new_session_per_task or False
 
         self._prof.prof('task_run_start', uid=tid)
-        task['proc'] = sp.Popen(args              = cmdline,
+        task['proc'] = sp.Popen(args              = launch_path,
                                 executable        = None,
                                 shell             = False,
                                 stdin             = None,
@@ -546,260 +384,6 @@ class Popen(AgentExecutingComponent):
                                                publish=True, push=True)
 
         return action
-
-
-    # --------------------------------------------------------------------------
-    #
-    # pylint: disable=unused-argument
-    def _get_prof(self, event, tid, msg=''):
-
-        return '$RP_PROF %s "%s"\n' % (event, msg)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _get_rp_funcs(self):
-
-        # define helper functions
-        ret  = '\nrp_error() {\n'
-        ret += '    echo "$1 failed" 1>&2\n'
-        ret += '    exit 1\n'
-        ret += '}\n'
-
-        return ret
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _get_rp_env(self, task):
-
-        tid  = task['uid']
-        td   = task['description']
-        name = task.get('name') or tid
-        sbox = os.path.realpath(task['task_sandbox_path'])
-
-        if sbox.startswith(self._pwd):
-            sbox = '$RP_PILOT_SANDBOX%s' % sbox[len(self._pwd):]
-
-        gpr = td['gpus_per_rank']
-        if int(gpr) == gpr:
-            gpr = '%d' % gpr
-        else:
-            gpr = '%f' % gpr
-
-        ret  = '\n'
-        ret += 'export RP_TASK_ID="%s"\n'          % tid
-        ret += 'export RP_TASK_NAME="%s"\n'        % name
-        ret += 'export RP_PILOT_ID="%s"\n'         % self._pid
-        ret += 'export RP_SESSION_ID="%s"\n'       % self.sid
-        ret += 'export RP_RESOURCE="%s"\n'         % self.resource
-        ret += 'export RP_RESOURCE_SANDBOX="%s"\n' % self.rsbox
-        ret += 'export RP_SESSION_SANDBOX="%s"\n'  % self.ssbox
-        ret += 'export RP_PILOT_SANDBOX="%s"\n'    % self.psbox
-        ret += 'export RP_TASK_SANDBOX="%s"\n'     % sbox
-        ret += 'export RP_REGISTRY_ADDRESS="%s"\n' % self.session.reg_addr
-        ret += 'export RP_CORES_PER_RANK=%d\n'     % td['cores_per_rank']
-        ret += 'export RP_GPUS_PER_RANK=%s\n'      % gpr
-
-        # FIXME AM
-      # ret += 'export RP_LFS="%s"\n'              % self.lfs
-        ret += 'export RP_GTOD="%s"\n'             % self.gtod
-        ret += 'export RP_PROF="%s"\n'             % self.prof
-
-        if self._prof.enabled:
-            ret += 'export RP_PROF_TGT="%s/%s.prof"\n' % (sbox, tid)
-        else:
-            ret += 'unset  RP_PROF_TGT\n'
-
-        return ret
-
-
-    # --------------------------------------------------------------------------
-    #
-    # launcher
-    #
-    def _get_launch_env(self, launcher):
-
-        ret  = ''
-
-        for cmd in launcher.get_launcher_env():
-            ret += '%s || rp_error launcher_env\n' % cmd
-
-        return ret
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _get_prep_launch(self, task, sig):
-
-        ret = ''
-        td = task['description']
-
-        if sig not in td:
-            return ret
-
-        for cmd in ru.as_list(task['description'][sig]):
-            ret += '%s || rp_error %s\n' % (cmd, sig)
-
-        return ret
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _get_launch(self, task, launcher, exec_path):
-
-        ret  = '( \\\n'
-
-        for cmd in ru.as_list(launcher.get_launch_cmds(task, exec_path)):
-            ret += '  %s \\\n' % cmd
-
-        ret += ') 1> %s \\\n  2> %s\n' % (task['stdout_file_short'],
-                                          task['stderr_file_short'])
-        # collect PID for launch-script
-        ret += 'RP_RET=$?\n'
-        ret += 'RP_LAUNCH_PID=$$\n'
-
-        return ret
-
-
-    # --------------------------------------------------------------------------
-    #
-    # exec
-    #
-    def _get_task_env(self, task, launcher):
-
-        ret = ''
-        td  = task['description']
-
-        # named_env's are prepared by the launcher
-        if td['named_env']:
-            ret += '\n# named environment\n'
-            ret += '. %s\n' % launcher.get_task_named_env(td['named_env'])
-
-        # also add any env vars requested in the task description
-        if td['environment']:
-            ret += '\n# task env settings\n'
-            for key, val in td['environment'].items():
-                ret += 'export %s="%s"\n' % (key, val)
-
-        return ret
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _get_rank_ids(self, n_ranks, launcher):
-
-        ret  = ''
-        ret += 'export RP_RANKS=%s\n' % n_ranks
-        ret += launcher.get_rank_cmd()
-
-        if n_ranks > 1:
-
-            # make sure that RP_RANK is known (otherwise task fails silently)
-            if 'export RP_RANK=' not in ret:
-                raise RuntimeError('launch method does not export RP_RANK')
-
-        # also define a method to sync all ranks on certain events
-        ret += '\nrp_sync_ranks() {\n'
-        ret += '    sig=$1\n'
-        ret += '    echo $RP_RANK >> $sig.sig\n'
-        ret += '    while test $(cat $sig.sig | wc -l) -lt $RP_RANKS; do\n'
-        ret += '        sleep 1\n'
-        ret += '    done\n'
-        ret += '}\n'
-
-        return ret
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _extend_pre_exec(self, td, ranks=None):
-
-        # FIXME: this assumes that the rank has a `gpu_maps` and `core_maps`
-        #        with exactly one entry, corresponding to the rank process to be
-        #        started.
-
-        # FIXME: need to distinguish between logical and physical IDs
-
-        if td['threading_type'] == rpc.OpenMP:
-            # for future updates: if task ranks are heterogeneous in terms of
-            #                     number of threads, then the following string
-            #                     should be converted into dictionary (per rank)
-            num_threads = td.get('cores_per_rank', 1)
-            td['pre_exec'].append('export OMP_NUM_THREADS=%d' % num_threads)
-
-        if td['gpus_per_rank'] and td['gpu_type'] == rpc.CUDA and ranks:
-            # equivalent to the 'physical' value for original `cvd_id_mode`
-            rank_id  = 0
-            rank_env = {}
-            for slot_ranks in ranks:
-                for gpu_map in slot_ranks['gpu_map']:
-                    rank_env[str(rank_id)] = \
-                        'export CUDA_VISIBLE_DEVICES=%s' % \
-                        ','.join([str(g) for g in gpu_map])
-                    rank_id += 1
-            td['pre_exec'].append(rank_env)
-
-        # pre-defined `pre_exec` per platform configuration
-        td['pre_exec'].extend(ru.as_list(self.session.rcfg.get('task_pre_exec')))
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _get_prep_exec(self, task, n_ranks, sig):
-
-        ret = ''
-        td  = task['description']
-
-        if sig not in td:
-            return ret
-
-        entries         = ru.as_list(td[sig])
-        switch_per_rank = any([isinstance(x, dict) for x in entries])
-        cmd_template    = '%s || rp_error %s\n'
-
-        sync_ranks_cmd = ''
-        if sig == 'pre_exec' and td['pre_exec_sync']:
-            sync_ranks_cmd = 'rp_sync_ranks %s\n' % sig
-
-        if not switch_per_rank:
-            return ''.join([cmd_template % (x, sig) for x in entries]) + \
-                   sync_ranks_cmd
-
-        ret += 'case "$RP_RANK" in\n'
-        for rank_id in range(n_ranks):
-
-            ret += '    %d)\n' % rank_id
-
-            for entry in entries:
-
-                if isinstance(entry, str):
-                    entry = {str(rank_id): entry}
-
-                for cmd in ru.as_list(entry.get(str(rank_id))):
-                    ret += '        ' + cmd_template % (cmd, sig)
-
-            ret += '        ;;\n'
-
-        ret += 'esac\n' + sync_ranks_cmd
-
-        return ret
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _get_exec(self, task, launcher):
-
-        # FIXME: core pinning goes here
-
-        ret  = '%s &\n' % launcher.get_exec(task)
-        # collect PIDs for exec-script and executable
-        ret += '\nRP_EXEC_PID=$$\nRP_RANK_PID=$!\n\n'
-        ret += 'wait $RP_RANK_PID\n'
-        # set output
-        ret += 'RP_RET=$?\n'
-
-        return ret
 
 
 # ------------------------------------------------------------------------------
