+++ src/radical/pilot/agent/launch_method/base.py
@@ -227,7 +227,13 @@ class LaunchMethod(object):
     #
     def get_exec(self, task):
 
-        raise NotImplementedError("incomplete LaunchMethod %s" % self.name)
+        td           = task['description']
+        task_exec    = td['executable']
+        task_args    = td['arguments']
+        task_argstr  = self._create_arg_string(task_args)
+        command      = '%s %s' % (task_exec, task_argstr)
+
+        return command.rstrip()
 
 
     # --------------------------------------------------------------------------
