+++ src/radical/pilot/agent/launch_method/mpiexec.py
@@ -293,18 +293,5 @@ class MPIExec(LaunchMethod):
         return ret
 
 
-    # --------------------------------------------------------------------------
-    #
-    def get_exec(self, task):
-
-        td           = task['description']
-        task_exec    = td['executable']
-        task_args    = td['arguments']
-        task_argstr  = self._create_arg_string(task_args)
-        command      = '%s %s' % (task_exec, task_argstr)
-
-        return command.rstrip()
-
-
 # ------------------------------------------------------------------------------
