+++ src/radical/pilot/agent/launch_method/mpirun.py
@@ -239,18 +239,5 @@ class MPIRun(LaunchMethod):
         return ret
 
 
-    # --------------------------------------------------------------------------
-    #
-    def get_exec(self, task):
-
-        td           = task['description']
-        task_exec    = td['executable']
-        task_args    = td['arguments']
-        task_argstr  = self._create_arg_string(task_args)
-        command      = '%s %s' % (task_exec, task_argstr)
-
-        return command.rstrip()
-
-
 # ------------------------------------------------------------------------------
