+++ src/radical/pilot/agent/launch_method/srun.py
@@ -202,18 +202,5 @@ class Srun(LaunchMethod):
         return ret
 
 
-    # --------------------------------------------------------------------------
-    #
-    def get_exec(self, task):
-
-        td          = task['description']
-        task_exec   = td['executable']
-        task_args   = td.get('arguments')
-        task_argstr = self._create_arg_string(task_args)
-        command     = '%s %s' % (task_exec, task_argstr)
-
-        return command.rstrip()
-
-
 # ------------------------------------------------------------------------------
